pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "linknower"
version = "0.2.0"
description = "Local-first personal workflow intelligence engine for developers"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "LinkNower Contributors"}
]
keywords = ["semantic-search", "workflow", "productivity", "local-first", "developer-tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Tools",
    "Topic :: Utilities",
]

dependencies = [
    "chromadb>=0.4.18",
    "sentence-transformers>=2.2.2",
    "hdbscan>=0.8.33",
    "umap-learn>=0.5.5",
    "typer[all]>=0.9.0",
    "rich>=13.7.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "gitpython>=3.1.40",
    "pyyaml>=6.0.1",
    "numpy>=1.24.0,<2.0.0",
    "streamlit>=1.28.0",
    "plotly>=5.18.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "isort>=5.13.0",
    "mypy>=1.7.0",
    "ruff>=0.1.8",
]

[project.scripts]
lk = "linknower.cli.main:app"

[project.urls]
Homepage = "https://github.com/linknower/linknower"
Documentation = "https://github.com/linknower/linknower#readme"
Repository = "https://github.com/linknower/linknower"
"Bug Tracker" = "https://github.com/linknower/linknower/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["linknower*"]
namespaces = false

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=linknower --cov-report=term-missing"
